new ver.

+++ tmp/arrays/MultiDeclForComponentExtensions13.kt
@@ -1,7 +1,7 @@
-operator fun Long.component1() = this + 1
-operator fun Long.component2() = this + 2
+operator fun Int.component1() = this + 1
+operator fun Int.component2() = this + 2
 
-fun doTest(l : Array<Long>): String {
+fun doTest(l : Array<Int>): String {
     var s = ""
     for ((a, b) in l) {
       s += "$a:$b;"
@@ -10,7 +10,7 @@ fun doTest(l : Array<Long>): String {
 }
 
 fun box(): String {
-  val l = Array<Long>(3, {x -> x.toLong()})
+  val l = Array<Int>(3, {x -> x})
   val s = doTest(l)
   return if (s == "1:2;2:3;3:4;") "OK" else "fail: $s"
 }
